requirements.txt
torch==2.9.1
torchvision==0.24.1
torchaudio==2.9.1

transformers==4.52.3
timm==0.5.4
ttach==0.0.3

numpy==2.2.1
opencv-python==4.12.0.88
Pillow==11.0.0
scipy==1.15.3
scikit-image==0.26.0

matplotlib==3.10.0
tqdm==4.67.1

packaging==24.2
PyYAML==6.0.2
